kernel: move time related code to new files gaptime.{c.h}
This also gets rid of some pointless calls to Panic: even if getrusage
fails, we can recover, so no reason to crash the whole program.
