+++ Makefile.rules
@@ -55,6 +55,7 @@ SOURCES += src/ffdata.c
 SOURCES += src/finfield.c
 SOURCES += src/funcs.c
 SOURCES += src/gap.c
+SOURCES += src/gaptime.c
 SOURCES += gen/gap_version.c	# generated source file
 SOURCES += src/gvars.c
 SOURCES += src/hookintrprtr.c
@@ -71,8 +72,8 @@ SOURCES += src/listfunc.c
 SOURCES += src/listoper.c
 SOURCES += src/lists.c
 SOURCES += src/macfloat.c
-SOURCES += src/modules.c
 SOURCES += src/modules_builtin.c
+SOURCES += src/modules.c
 SOURCES += src/objcftl.c
 SOURCES += src/objects.c
 SOURCES += src/objfgelm.cc
@@ -99,10 +100,9 @@ SOURCES += src/syntaxtree.c
 SOURCES += src/sysfiles.c
 SOURCES += src/sysjmp.c
 SOURCES += src/sysmem.c
-SOURCES += src/sysstr.c
 SOURCES += src/sysroots.c
+SOURCES += src/sysstr.c
 SOURCES += src/system.c
-SOURCES += src/systime.c
 SOURCES += src/tietze.c
 SOURCES += src/tracing.c
 SOURCES += src/trans.cc
