+++ src/sysmem.c
@@ -14,10 +14,10 @@
 
 #include "sysmem.h"
 
+#include "gaptime.h"
 #include "stats.h"
 #include "sysfiles.h"
 #include "sysopt.h"
-#include "systime.h"
 
 #ifdef GAP_MEM_CHECK
 #include <fcntl.h>
